requirements.txt
tensorflow==2.15.0
streamlit
matplotlib
pillow
numpy